feat(monorepo): update release tooling, to aid migration to automated releases
