+++ package.json
@@ -70,7 +70,8 @@
     "full-build": "npm i && npm run symlinks && npm run build:code && npm run build:sample && npm run build:docs && npm run jest:test && npm run cypress:e2e:server",
     "workspace-generator": "nx workspace-generator",
     "auto-publish-nightly": "tsc -p tools/tsconfig.tools.json && node dist/tools/autoVersion",
-    "release:npm": "tsc -p tools/tsconfig.tools.json && node dist/tools/deployReleasets.js",
+    "prepRelease": "tsc -p tools/tsconfig.tools.json && cp tools/package.json dist/tools/. && node dist/tools/tagRelease.js",
+    "doPublish": "tsc -p tools/tsconfig.tools.json && cp tools/package.json dist/tools/. && node dist/tools/publishCurrent.js",
     "prepare": "husky install"
   },
   "private": true,
