+++ tools/publishPackage.ts
@@ -1,9 +1,13 @@
-import { green, yellow } from 'chalk';
 import { exec } from 'child_process';
+import { createRequire } from 'module';
 import { join } from 'path';
-import { folder, ReleaseData } from './utils';
+import { folder, ReleaseData } from './utils.js';
+
+const require= createRequire(import.meta.url);
+const { green, yellow } = require('chalk');
 
 export async function publishPackage(tag: string, toRelease: ReleaseData, dryRun = true) {
+
   const cmd = `npm publish --access public --ignore-scripts --tag ${tag}` + (dryRun ? ' --dry-run' : '');
   const { nodeError, stdOut, stdErr } = await new Promise((resolve, reject) => {
     // resolve({
