+++ tools/tsconfig.tools.json
@@ -2,11 +2,11 @@
   "compilerOptions": {
     "outDir": "../dist/tools",
     "rootDir": ".",
-    "target": "es2020",
+    "target": "ES2020",
     "moduleResolution": "Node",
-    "module": "CommonJS",
-    "lib": ["ES2020"],
+    "module": "ESNext",
+    "lib": ["es2021"],
     "types": ["node"]
   },
-  "include": ["**/*.ts"]
+  "include": ["**/*.ts", "./deployRelease.mts"]
 }
